Added MenuWindow and PopularCrypto Classes
MenuWindow will be the initial JFrame which will popup to the user when they first run the program. It will have 2 buttons to view the top 10 crypto (by market cap) or the most popular crypto according to retail investors.

PopularCrypto will be the JFrame window which will display the content in regards to the most popular crypto according to retail investors/media.

Renamed CryptoPriceIndicator class to TopTenCrypto